+++ Source/Project_Lyzandred/LevelLoader/LevelDataAsset.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Engine/DataAsset.h"
+#include "LevelDataAsset.generated.h"
+
+UCLASS(BlueprintType)
+class PROJECT_LYZANDRED_API ULevelDataAsset : public UDataAsset
+{
+	GENERATED_BODY()
+	
+public:
+	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category="Level")
+	FName LevelName = FName("Level");
+
+	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category="Level")
+	TSet<TSoftObjectPtr<UWorld>> Sublevels;
+};