+++ Source/Project_Lyzandred/Other/SubsystemConfig.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Engine/DataAsset.h"
+#include "SubsystemConfig.generated.h"
+
+UCLASS(Abstract, BlueprintType)
+class PROJECT_LYZANDRED_API USubsystemConfig : public UDataAsset
+{
+	GENERATED_BODY()
+	
+};